% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{null}
\alias{null}
\title{Null function}
\usage{
null(x = NULL)
}
\arguments{
\item{x}{(default, NULL)}
}
\description{
Use this as template for your first function, and delete it!
}
\note{
You can add all this documentation infrastructure pressing
\code{CTRL + SHIFT + ALT + R} from anywhere inside the function's body.
}
\examples{
\dontrun{
  null()
  null(1)
}
}
